requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
coverage>=7.3.0
unittest-xml-reporting>=3.2.0
moto[all]>=4.2.14
freezegun>=1.2.2
localstack>=2.3.0
boto3-stubs[essential]>=1.34.0